SummarizationModule/backend/requirements.txt
Flask>=3.1,<4
flask-cors>=5.0,<6
portkey-ai>=1.9,<2
pydantic>=2.10,<3
openpyxl>=3.1,<4
reportlab>=4.2,<5
Pillow>=11,<12
python-dotenv>=1.0,<2
pandas>=2.2,<3
httpx>=0.28,<1
